requirements.txt
torch
transformers==4.25.1
spikingjelly
torchvision
tqdm
numpy
pillow
scikit-learn
PyYAML